Started an "Add polygon" dialog.
The AddPolygonDialog is added, it's appearance and behavior is
approached to the old one.

Refactoring: old classes are renamed to start with capital letter;
defines from RenderUtilities.h are moved to the header.

Started an "Add polygon" dialog.

The AddPolygonDialog is added, it's appearance and behavior is
approached to the old one.

Refactoring: old classes are renamed to start with capital letter;
defines from RenderUtilities.h are moved to the header.